requirements.txt
aws-cdk-lib==2.171.1
pydantic==2.7.0
pyaml==24.4.0

